Structure of the routing mostly done... Misses few things.
