+++ lib/root/root_view.dart
@@ -1,0 +1,27 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_app/store/router.dart';
+import 'package:flutter_app/store/store.dart';
+import 'package:flutter_redux/flutter_redux.dart';
+
+import 'package:redux/redux.dart';
+
+class RootWidget extends StatelessWidget {
+  final store = Store((x, a) => x, initialState: AppState.init());
+  RootWidget({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return StoreProvider(store: store, child: _RouterWidget());
+  }
+
+}
+
+class _RouterWidget extends StatelessWidget {
+
+  @override
+  Widget build(BuildContext context) {
+    return StoreConnector<AppState, RouterA>(builder: (context, router) {
+      return router.getWidget();
+    }, converter: (store) => store.state.router);
+  }
+}
