+++ lib/store/store.dart
@@ -1,6 +1,8 @@
 
 import 'package:flutter/material.dart';
 import 'package:flutter_app/login/store/store.dart';
+import 'package:flutter_app/root/store/route.dart';
+import 'package:flutter_app/root/store/store.dart';
 import 'package:flutter_app/store/router.dart';
 import 'package:intl/date_symbols.dart';
 
@@ -8,10 +10,11 @@ import 'package:intl/date_symbols.dart';
 class AppState {
   final RouterA router;
   final LoginState loginViewState;
+  final PageControlState pageControlState;
   // add User, ...
-  AppState(this.router, this.loginViewState);
+  AppState(this.router, this.loginViewState, this.pageControlState);
 
-  AppState.init() : router = TODO, loginViewState = LoginState.init();
+  AppState.init() : router = PageControlRouter(), loginViewState = LoginState.init(), pageControlState = PageControlState.init();
 }
